[#283] refactor: badge로 package 모으는 걸로 변경.
