+++ src/main/java/goorm/eagle7/stelligence/domain/badge/event/listener/NewContributeEventListener.java
@@ -1,4 +1,4 @@
-package goorm.eagle7.stelligence.domain.contribute.event.listener;
+package goorm.eagle7.stelligence.domain.badge.event.listener;
 
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
@@ -9,26 +9,26 @@ import org.springframework.transaction.event.TransactionalEventListener;
 import goorm.eagle7.stelligence.domain.badge.BadgeService;
 import goorm.eagle7.stelligence.domain.badge.model.BadgeCategory;
 import goorm.eagle7.stelligence.domain.contribute.ContributeRepository;
-import goorm.eagle7.stelligence.domain.contribute.event.ContributeRejectedEvent;
+import goorm.eagle7.stelligence.domain.contribute.event.NewContributeEvent;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class ContributeRejectedEventListener {
+public class NewContributeEventListener {
 
 	private final BadgeService badgeService;
 	private final ContributeRepository contributeRepository;
 
 	@Async
 	@Transactional(propagation = Propagation.REQUIRES_NEW)
-	@TransactionalEventListener(value = ContributeRejectedEvent.class)
-	public void onContributeRejected(ContributeRejectedEvent event) {
+	@TransactionalEventListener(value = NewContributeEvent.class)
+	public void onContributeNew(NewContributeEvent event) {
 
 		contributeRepository
 			.findWithMember(event.contributeId())
 			.ifPresent(contribute ->
 				badgeService.checkAndAwardBadge(
-					BadgeCategory.CONTRIBUTE_REJECTED
+					BadgeCategory.CONTRIBUTE_ALL
 					, contribute.getMember())
 			);
